backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyJWT
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
python-multipart
